requirements.txt
pyTelegramBotAPI==4.14.0
supabase==2.0.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0